+++ spec/lib/tmuxinator/project_spec.rb
@@ -665,7 +665,7 @@ describe Tmuxinator::Project do
   end
 
   describe "::parse_settings" do
-    let(:args) { ["one", "two=three"] }
+    let(:args) { ["one", "two=three", "four=five=six"] }
 
     it "returns settings in a hash" do
       expect(described_class.parse_settings(args)["two"]).to eq("three")
@@ -674,6 +674,10 @@ describe Tmuxinator::Project do
     it "removes settings from args" do
       described_class.parse_settings(args)
       expect(args).to eq(["one"])
+    end
+
+    it "handles equals signs in values" do
+      expect(described_class.parse_settings(args)["four"]).to eq("five=six")
     end
   end